test/shared/mongo: merge uri.go into client.go
NewFromURI is a one-function helper that overlaps heavily with

NewClient (both call mongo.Connect on an ApplyURI options chain

with DefaultConnectTimeout). Folding it into client.go removes a

duplicate package doc comment and keeps both constructors and

their tests next to each other.

(cherry picked from commit 29ab00b30b302345da8b533dca49288f4ace451c)
