+++ test/shared/mongo/client.go
@@ -123,6 +123,27 @@ func NewClient(_ context.Context, opts ClientOptions) (*mongo.Client, error) {
 	return c, nil
 }
 
+// NewFromURI builds a connected *mongo.Client against an externally
+// supplied mongodb:// URI. Connect time is bounded by
+// DefaultConnectTimeout. The driver's Connect is lazy, so callers who
+// need a post-connect round-trip should call Ping themselves.
+//
+// This helper exists so the long-haul driver can stop calling
+// mongo.Connect(options.Client().ApplyURI(...)) directly, which would
+// silently bypass the connect timeout that NewClient applies for
+// every other caller.
+func NewFromURI(_ context.Context, uri string) (*mongo.Client, error) {
+	if uri == "" {
+		return nil, errors.New("mongo: uri is required")
+	}
+	co := options.Client().ApplyURI(uri).SetConnectTimeout(DefaultConnectTimeout)
+	c, err := mongo.Connect(co)
+	if err != nil {
+		return nil, fmt.Errorf("mongo: connect: %w", err)
+	}
+	return c, nil
+}
+
 // buildTLSConfig assembles a *tls.Config for the driver. Priority:
 //
 //  1. RootCAs, if non-nil — use as trust store.
